Fix MSVC compile issues caused by my changes

+++ dtc/src/emitter/LabelMapping.cpp
@@ -21,7 +21,7 @@ uint32_t LabelMapping::add_label(const string &label) noexcept {
     labels.insert(make_pair(label, strings_pos));
     uint32_t address = strings_pos;
 
-    strings_pos += label.size() + 1;
+    strings_pos += (uint32_t)label.size() + 1;
 
     return address;
 }
@@ -29,7 +29,7 @@ uint32_t LabelMapping::add_label(const string &label) noexcept {
 void LabelMapping::write_labels(DTBFile &file) const noexcept {
     uint32_t count;
     size_t pos = 0;
-    for(count = labels.size(); count > 0; count--) {
+    for(count = (uint32_t)labels.size(); count > 0; count--) {
         for(auto const& x : labels) {
             if (x.second == pos) {
                 file.write(x.first);
